Modificación en los controladores del modelo MVC, ordenando y reutilizando las listas del controlador índex
