+++ src/main/java/com/example/stose/controller/indexController.java
@@ -10,7 +10,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.example.stose.model.indexModel;
-import com.example.stose.model.tiendaModel;
 
 
 
@@ -76,7 +75,7 @@ public class indexController {
 
 				//CONTENIDO DEL FOOTER
 				//Ayuda y contacto
-				tiendaModel AyudaC = new tiendaModel();
+				indexModel AyudaC = new indexModel();
 				AyudaC.setCuenta("Tu Cuenta");
 				AyudaC.setPedidos("Tus Pedidos");
 				AyudaC.setEnvios("Tarifas de Envio");
@@ -84,11 +83,11 @@ public class indexController {
 				AyudaC.setContactanos("Contactanos");
 				
 				
-				List<tiendaModel> ListAyudaC = new ArrayList<>();
+				List<indexModel> ListAyudaC = new ArrayList<>();
 				ListAyudaC.add(AyudaC);
 
 		        //Libros y categorias
-				tiendaModel LibrosC = new tiendaModel();
+				indexModel LibrosC = new indexModel();
 				LibrosC.setArte("Arte");
 				LibrosC.setBiografias("Biografias");
 				LibrosC.setCiencias("Ciencias");
@@ -101,26 +100,26 @@ public class indexController {
 				LibrosC.setFilosofiaYreligion("Filosofia y Religion");
 				LibrosC.setInfantilesYjuveniles("Infantiles y Juveniles");
 
-		        List<tiendaModel> ListLibrosC = new ArrayList<>();
+		        List<indexModel> ListLibrosC = new ArrayList<>();
 				ListLibrosC.add(LibrosC);
 
 				//Metodos de pago
-				tiendaModel MetodosP = new tiendaModel();
+				indexModel MetodosP = new indexModel();
 				MetodosP.setTarjetasDeCredito("Tarjetas de Credito");
 				MetodosP.setPaypal("Paypal");
 				
-				List<tiendaModel> ListMetodosP = new ArrayList<>();
+				List<indexModel> ListMetodosP = new ArrayList<>();
 				ListMetodosP.add(MetodosP);
 
 				//Sobre Nosotros
-				tiendaModel Nosotros = new tiendaModel();
+				indexModel Nosotros = new indexModel();
 				Nosotros.setIempresa("Informacion de la Compañia");
 				Nosotros.setNoticias("Noticias");
 				Nosotros.setInversores("Inversores");
 				Nosotros.setPoliticas("Politicas");
 				Nosotros.setClienteV("Valoracion de los clientes");
 				
-				List<tiendaModel> ListNosotros = new ArrayList<>();
+				List<indexModel> ListNosotros = new ArrayList<>();
 				ListNosotros.add(Nosotros);
 
 		        //Models
