+++ src/main/java/com/example/stose/model/indexModel.java
@@ -30,6 +30,54 @@ public class indexModel {
 	    private String LibroDestacado7;
 	    private String LibroDestacado8;
 	    
+	  //Ayuda y contacto
+	  		private String Cuenta;
+	  		private String Pedidos;
+	  		private String Envios;
+	  		private String Devoluciones;
+	  		private String Contactanos;
+	  		
+	  		//Libros y Categorias
+	  		private String Arte;
+	  		private String Biografias;
+	  		private String Ciencias;
+	  		private String Computacion;
+	  		private String Deportes;
+	  		private String Derecho;
+	  		private String Economia;
+	  		private String EstilosDeVida;
+	  		private String Ficcion;
+	  		private String FilosofiaYreligion;
+	  		private String InfantilesYjuveniles;
+	  		
+	  		//Metodos de pagos
+	  		private String TarjetasDeCredito;
+	  		private String Paypal;
+	  		
+	  		//Sobre nosotros
+	  		
+	  		private String Iempresa;
+	  		private String Noticias;
+	  		private String Inversores;
+	  		private String Politicas;
+	  		private String ClienteV;
+	  		
+	  		//Header top
+	  		
+	  		private String Blog;
+	  		private String ProductosDestacados;
+	  		private String ListaDeDeseos;
+
+	  		private String Registrate;
+	  		private String MiCuenta;
+	  		private String RevisionDePedidos;
+	  		
+	  		//Header nav
+	  		
+	  		private String Inicio;
+	  		private String Tienda;
+	  		private String CostoDeEnvio;
+	  		private String Nosotros;
 	    //CARRITO DE COMPRAS
 	    
 	    public String getCarritoCompras1() {
@@ -168,4 +216,221 @@ public class indexModel {
 	    public void setLibroDestacado8(String libroDestacado8) {
 	        LibroDestacado8 = libroDestacado8;
 	    }
+	  //HEADER NAV
+	  		public String getInicio() {
+	  			return Inicio;
+	  		}
+	  		public void setInicio(String inicio) {
+	  			Inicio = inicio;
+	  		}
+	  		public String getTienda() {
+	  			return Tienda;
+	  		}
+	  		public void setTienda(String tienda) {
+	  			Tienda = tienda;
+	  		}
+	  		public String getCostoDeEnvio() {
+	  			return CostoDeEnvio;
+	  		}
+	  		public void setCostoDeEnvio(String costoDeEnvio) {
+	  			CostoDeEnvio = costoDeEnvio;
+	  		}
+	  		public String getNosotros() {
+	  			return Nosotros;
+	  		}
+	  		public void setNosotros(String nosotros) {
+	  			Nosotros = nosotros;
+	  		}
+	  		
+	  		
+	  		//HEADER TOP
+	  		public String getBlog() {
+	  			return Blog;
+	  		}
+	  		public void setBlog(String blog) {
+	  			Blog = blog;
+	  		}
+	  		public String getProductosDestacados() {
+	  			return ProductosDestacados;
+	  		}
+	  		public void setProductosDestacados(String productosDestacados) {
+	  			ProductosDestacados = productosDestacados;
+	  		}
+	  		public String getListaDeDeseos() {
+	  			return ListaDeDeseos;
+	  		}
+	  		public void setListaDeDeseos(String listaDeDeseos) {
+	  			ListaDeDeseos = listaDeDeseos;
+	  		}
+	  		public String getRegistrate() {
+	  			return Registrate;
+	  		}
+	  		public void setRegistrate(String registrate) {
+	  			Registrate = registrate;
+	  		}
+	  		public String getMiCuenta() {
+	  			return MiCuenta;
+	  		}
+	  		public void setMiCuenta(String miCuenta) {
+	  			MiCuenta = miCuenta;
+	  		}
+	  		public String getRevisionDePedidos() {
+	  			return RevisionDePedidos;
+	  		}
+	  		public void setRevisionDePedidos(String revisionDePedidos) {
+	  			RevisionDePedidos = revisionDePedidos;
+	  		}
+	  		//SOBRE NOSOTROS
+	  		public String getIempresa() {
+	  			return Iempresa;
+	  		}
+	  		public void setIempresa(String iempresa) {
+	  			Iempresa = iempresa;
+	  		}
+	  		public String getNoticias() {
+	  			return Noticias;
+	  		}
+	  		public void setNoticias(String noticias) {
+	  			Noticias = noticias;
+	  		}
+	  		public String getInversores() {
+	  			return Inversores;
+	  		}
+	  		public void setInversores(String inversores) {
+	  			Inversores = inversores;
+	  		}
+	  		public String getPoliticas() {
+	  			return Politicas;
+	  		}
+	  		public void setPoliticas(String politicas) {
+	  			Politicas = politicas;
+	  		}
+	  		public String getClienteV() {
+	  			return ClienteV;
+	  		}
+	  		public void setClienteV(String clienteV) {
+	  			ClienteV = clienteV;
+	  		}
+	  		
+	  		
+	  		
+	  		
+	  		
+	  		//METODOS DE PAGO
+	  		public String getTarjetasDeCredito() {
+	  			return TarjetasDeCredito;
+	  		}
+	  		public void setTarjetasDeCredito(String tarjetasDeCredito) {
+	  			TarjetasDeCredito = tarjetasDeCredito;
+	  		}
+	  		public String getPaypal() {
+	  			return Paypal;
+	  		}
+	  		public void setPaypal(String paypal) {
+	  			Paypal = paypal;
+	  		}
+	  		
+	  		
+	  		
+	  		
+	  		//LIBROS Y CATEGORIAS
+	  		public String getArte() {
+	  			return Arte;
+	  		}
+	  		public void setArte(String arte) {
+	  			Arte = arte;
+	  		}
+	  		public String getBiografias() {
+	  			return Biografias;
+	  		}
+	  		public void setBiografias(String biografias) {
+	  			Biografias = biografias;
+	  		}
+	  		public String getCiencias() {
+	  			return Ciencias;
+	  		}
+	  		public void setCiencias(String ciencias) {
+	  			Ciencias = ciencias;
+	  		}
+	  		public String getComputacion() {
+	  			return Computacion;
+	  		}
+	  		public void setComputacion(String computacion) {
+	  			Computacion = computacion;
+	  		}
+	  		public String getDeportes() {
+	  			return Deportes;
+	  		}
+	  		public void setDeportes(String deportes) {
+	  			Deportes = deportes;
+	  		}
+	  		public String getDerecho() {
+	  			return Derecho;
+	  		}
+	  		public void setDerecho(String derecho) {
+	  			Derecho = derecho;
+	  		}
+	  		public String getEconomia() {
+	  			return Economia;
+	  		}
+	  		public void setEconomia(String economia) {
+	  			Economia = economia;
+	  		}
+	  		public String getEstilosDeVida() {
+	  			return EstilosDeVida;
+	  		}
+	  		public void setEstilosDeVida(String estilosDeVida) {
+	  			EstilosDeVida = estilosDeVida;
+	  		}
+	  		public String getFiccion() {
+	  			return Ficcion;
+	  		}
+	  		public void setFiccion(String ficcion) {
+	  			Ficcion = ficcion;
+	  		}
+	  		public String getFilosofiaYreligion() {
+	  			return FilosofiaYreligion;
+	  		}
+	  		public void setFilosofiaYreligion(String filosofiaYreligion) {
+	  			FilosofiaYreligion = filosofiaYreligion;
+	  		}
+	  		public String getInfantilesYjuveniles() {
+	  			return InfantilesYjuveniles;
+	  		}
+	  		public void setInfantilesYjuveniles(String infantilesYjuveniles) {
+	  			InfantilesYjuveniles = infantilesYjuveniles;
+	  		}
+	  		
+	  		
+	  		//AYUDA Y CONTACTO
+	  		public String getCuenta() {
+	  			return Cuenta;
+	  		}
+	  		public void setCuenta(String cuenta) {
+	  			Cuenta = cuenta;
+	  		}
+	  		public String getPedidos() {
+	  			return Pedidos;
+	  		}
+	  		public void setPedidos(String pedidos) {
+	  			Pedidos = pedidos;
+	  		}
+	  		public String getEnvios() {
+	  			return Envios;
+	  		}
+	  		public void setEnvios(String envios) {
+	  			Envios = envios;
+	  		}
+	  		public String getDevoluciones() {
+	  			return Devoluciones;
+	  		}
+	  		public void setDevoluciones(String devoluciones) {
+	  			Devoluciones = devoluciones;
+	  		}
+	  		public String getContactanos() {
+	  			return Contactanos;
+	  		}
+	  		public void setContactanos(String contactanos) {
+	  			Contactanos = contactanos;
+	  		}
 }
