+++ src/main/java/com/example/stose/model/libroOfertaModel.java
@@ -1,11 +1,6 @@
 package com.example.stose.model;
 
 public class libroOfertaModel {
-	//Carrito de compras
-		private String CarritoCompras1;
-		private String CarritoCompras2;
-		private String CarritoCompras3;
-		private String CarritoCompras4;
 
 		//Detalle Libro
 		private String Titulo;
@@ -21,46 +16,7 @@ public class libroOfertaModel {
 		private String Editorial;
 		private String NumeroPaginas;
 		
-		//Libros relacionados
-		private String LibroRelacionado1;
-		private String LibroRelacionado2;
-		private String LibroRelacionado3;
-		private String LibroRelacionado4;
-	    private String LibroRelacionado5;
-		private String LibroRelacionado6;
-		private String LibroRelacionado7;
-		private String LibroRelacionado8;
 		
-
-	    //CARRITO DE COMPRAS
-	    
-	    public String getCarritoCompras1() {
-	        return CarritoCompras1;
-	    }
-	    public void setCarritoCompras1(String carritoCompras1) {
-	        CarritoCompras1 = carritoCompras1;
-	    }
-	    
-	    public String getCarritoCompras2() {
-	        return CarritoCompras2;
-	    }
-	    public void setCarritoCompras2(String carritoCompras2) {
-	        CarritoCompras2 = carritoCompras2;
-	    }
-	    
-	    public String getCarritoCompras3() {
-	        return CarritoCompras3;
-	    }
-	    public void setCarritoCompras3(String carritoCompras3) {
-	        CarritoCompras3 = carritoCompras3;
-	    }
-	    
-	    public String getCarritoCompras4() {
-	        return CarritoCompras4;
-	    }
-	    public void setCarritoCompras4(String carritoCompras4) {
-	        CarritoCompras4 = carritoCompras4;
-	    }
 		
 		//DETALLE LIBRO
 		public String getTitulo() {
@@ -130,60 +86,4 @@ public class libroOfertaModel {
 	    }
 		
 		
-		//LIBROS RELACIONADOS
-		public String getLibroRelacionado1() {
-	        return LibroRelacionado1;
-	    }
-	    public void setLibroRelacionado1(String libroRelacionado1) {
-	        LibroRelacionado1 = libroRelacionado1;
-	    }
-	    
-	    public String getLibroRelacionado2() {
-	        return LibroRelacionado2;
-	    }
-	    public void setLibroRelacionado2(String libroRelacionado2) {
-	        LibroRelacionado2 = libroRelacionado2;
-	    }
-	    
-	    public String getLibroRelacionado3() {
-	        return LibroRelacionado3;
-	    }
-	    public void setLibroRelacionado3(String libroRelacionado3) {
-	        LibroRelacionado3 = libroRelacionado3;
-	    }
-	    
-	    public String getLibroRelacionado4() {
-	        return LibroRelacionado4;
-	    }
-	    public void setLibroRelacionado4(String libroRelacionado4) {
-	        LibroRelacionado4 = libroRelacionado4;
-	    }
-	    
-	    public String getLibroRelacionado5() {
-	        return LibroRelacionado5;
-	    }
-	    public void setLibroRelacionado5(String libroRelacionado5) {
-	        LibroRelacionado5 = libroRelacionado5;
-	    }
-	    
-	    public String getLibroRelacionado6() {
-	        return LibroRelacionado6;
-	    }
-	    public void setLibroRelacionado6(String libroRelacionado6) {
-	        LibroRelacionado6 = libroRelacionado6;
-	    }
-	    
-	    public String getLibroRelacionado7() {
-	        return LibroRelacionado7;
-	    }
-	    public void setLibroRelacionado7(String libroRelacionado7) {
-	        LibroRelacionado7 = libroRelacionado7;
-	    }
-	    
-	    public String getLibroRelacionado8() {
-	        return LibroRelacionado8;
-	    }
-	    public void setLibroRelacionado8(String libroRelacionado8) {
-	        LibroRelacionado8 = libroRelacionado8;
-	    }
 }
